requirements.txt
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.1.7
contourpy==1.1.0
convertdate==2.4.0
cycler==0.11.0
Cython==3.0.2
ephem==4.1.4
filelock==3.12.2
Flask==2.2.5
fonttools==4.42.1
fsspec==2023.6.0
gunicorn==20.1.0
holidays==0.31
idna==3.4
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
LunarCalendar==0.0.9
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.26.1
onnx==1.15.0
openai==1.16.2
openai-multi-tool-use-parallel-patch==0.2.0
optimum==1.16.2
packaging==23.1
pandas==2.0.3
patsy==0.5.3
Pillow==10.0.0
PyGithub==2.1.1
PyMeeus==0.5.12
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
sentence_transformers==2.3.1
sentry-sdk @ git+https://github.com/getsentry/sentry-python.git@c675239ced6927ebefae8305c2c7b1d310079f0a
simdkalman==1.0.2
six==1.16.0
statsmodels==0.14.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.0.1
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
tree_sitter_languages==1.10.2
urllib3==1.26.16
Werkzeug==2.3.7
pre-commit==3.5.0
black==23.10.1
isort==5.12.0
flake8==6.1.0
pytest==7.4.3
pydantic==2.6.2
mypy==1.8.0
mypy-extensions==1.0.0
openapi-core==0.18.2
pandas-stubs==2.1.4.231227
types-Pillow==10.1.0
types-colorama==0.4.15.12
types-pywin32==306.0.0.8
types-setuptools==69.0.0.0
types-tabulate==0.9.0.3
types-tqdm==4.66.0.5
types_pytz==2023.3.1.1
types-openpyxl==3.1.0.32
types-PyYAML==6.0.12.1
types-jsonschema==4.20.0.20240105
types-requests==2.25.0
openapi-pydantic==0.4.0
celery==5.3.6
celery-stubs==0.1.3
redis==5.0.1
# deepsparse-nightly==1.7.0.20240103 TODO - fix pydantic conflict
unidiff==0.7.5
transformers==4.37.2
psycopg==3.1.18
pgvector==0.2.4
sqlalchemy==2.0.25
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
types-Flask-Migrate==4.0.0.20240205
langsmith==0.0.87
pytest-asyncio==0.23.5
aiohttp==3.9.3
types-python-dateutil==2.8.19.20240106
johen==0.1.3
pydantic-xml==2.9.0
chromadb==0.4.14
google-cloud-storage==2.16.0
